+++ example/src/pages/http/http.html
@@ -26,6 +26,9 @@
   <button (click)="getCookies()" ion-button>Get Cookies</button>
   <button (click)="testCookies()" ion-button>Test Cookies</button>
 
+  <button (click)="uploadFile()" ion-button>Upload File</button>
+  <button (click)="downloadFile()" ion-button>Download File</button>
+
   <h4>Output</h4>
   <ion-textarea [value]="output" placeholder="Ouput"></ion-textarea>
 </ion-content>
