+++ ios/Capacitor/Capacitor/Plugins/DefaultPlugins.m
@@ -77,6 +77,8 @@ CAP_PLUGIN(CAPHttpPlugin, "Http",
   CAP_PLUGIN_METHOD(request, CAPPluginReturnPromise);
   CAP_PLUGIN_METHOD(setCookie, CAPPluginReturnPromise);
   CAP_PLUGIN_METHOD(getCookies, CAPPluginReturnPromise);
+  CAP_PLUGIN_METHOD(downloadFile, CAPPluginReturnPromise);
+  CAP_PLUGIN_METHOD(uploadFile, CAPPluginReturnPromise);
 )
 
 CAP_PLUGIN(CAPKeyboard, "Keyboard",
